@@ -1,60 +1,86 @@
 {
-  "name": "red_tetrisboilerplate",
-  "version": "0.0.1",
-  "author": "redpelicans",
-  "license": "MIT",
-  "scripts": {
-    "eslint": "eslint src",
-    "client-dist": "NODE_ENV=production webpack --progress --colors",
-    "client-dev": "webpack-dev-server --colors --hot --inline --host 0.0.0.0 --port 8080",
-    "srv-dev": "DEBUG=tetris:* babel-watch -w src src/server/main.js",
-    "srv-dist": "DEBUG=tetris:* babel src --out-dir dist",
-    "test": "mocha \"test/**/*.js\" --require babel-core/register --reporter spec",
-    "coverage": "NODE_ENV=test nyc --include \"src/shared/**/*.js\" --include \"src/server/**/*.js\" --all --reporter=text --reporter=lcov --require babel-core/register mocha \"test/**/*.js\""
-
-  },
-  "dependencies": {
-    "react": "^15.0.2",
-    "react-dom": "^15.0.2",
-    "react-redux": "^4.4.5",
-    "redux": "^3.5.2",
-    "redux-logger": "^2.6.1",
-    "redux-thunk": "^2.0.1",
-    "socket.io": "^1.4.6"
-  },
-  "devDependencies": {
-    "babel-cli": "^6.7.7",
-    "babel-core": "^6.7.7",
-    "babel-eslint": "^6.1.0",
-    "babel-loader": "^6.2.4",
-    "babel-plugin-__coverage__": "^0.111111.1",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "babel-watch": "^2.0.2",
-    "chai": "^3.5.0",
-    "chai-as-promised": "^5.3.0",
-    "chai-equal-jsx": "^1.0.9",
-    "eslint": "^2.13.1",
-    "eslint-plugin-babel": "^3.3.0",
-    "eslint-plugin-react": "^5.2.2",
-    "mocha": "^2.4.5",
-    "nyc": "^6.4.4",
-    "webpack": "^1.13.0",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-dev-server": "^1.14.1",
-    "webpack-hot-middleware": "^2.10.0"
-  },
-  "nyc": {
-    "exclude": [
-      "build"
-    ]
-  },
-  "babel": {
-    "presets": [
-      "es2015",
-      "react",
-      "stage-0"
-    ]
-  }
+    "name": "multitetris",
+    "version": "0.0.1",
+    "type": "module",
+    "author": "lorene-k",
+    "license": "MIT",
+    "scripts": {
+        "eslint": "eslint src",
+        "client-dist": "NODE_ENV=production webpack --progress",
+        "client-dev": "webpack-dev-server --hot --host 0.0.0.0 --port 8080",
+        "srv-dev": "DEBUG=tetris:* nodemon --exec babel-node src/server/main.js",
+        "srv-dist": "DEBUG=tetris:* babel src --out-dir dist",
+        "test": "mocha \"test/**/*.js\" --require @babel/register --reporter spec",
+        "coverage": "NODE_ENV=test nyc --include \"src/shared/**/*.js\" --include \"src/server/**/*.js\" --all --reporter=text --reporter=lcov --require @babel/register mocha \"test/**/*.js\""
+    },
+    "dependencies": {
+        "@reduxjs/toolkit": "^2.11.2",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
+        "react-redux": "^9.2.0",
+        "redux": "^5.0.1",
+        "redux-logger": "^3.0.6",
+        "redux-thunk": "^3.1.0",
+        "seedrandom": "^3.0.5",
+        "socket.io": "^4.8.1",
+        "socket.io-client": "^4.8.1"
+    },
+    "devDependencies": {
+        "@babel/cli": "^7.25.9",
+        "@babel/core": "^7.26.0",
+        "@babel/node": "^7.26.0",
+        "@babel/preset-env": "^7.26.0",
+        "@babel/preset-react": "^7.25.9",
+        "@babel/register": "^7.25.9",
+        "@eslint/js": "^9.17.0",
+        "babel-loader": "^9.2.1",
+        "babel-plugin-istanbul": "^7.0.0",
+        "chai": "^4.5.0",
+        "chai-as-promised": "^8.0.0",
+        "eslint": "^9.17.0",
+        "eslint-plugin-react": "^7.37.5",
+        "globals": "^15.14.0",
+        "mocha": "^10.8.2",
+        "nodemon": "^3.1.9",
+        "nyc": "^17.1.0",
+        "webpack": "^5.104.1",
+        "webpack-cli": "^5.1.4",
+        "webpack-dev-server": "^5.2.2"
+    },
+    "babel": {
+        "presets": [
+            [
+                "@babel/preset-env",
+                {
+                    "targets": {
+                        "node": "current"
+                    },
+                    "modules": false
+                }
+            ],
+            [
+                "@babel/preset-react",
+                {
+                    "runtime": "automatic"
+                }
+            ]
+        ],
+        "env": {
+            "test": {
+                "presets": [
+                    [
+                        "@babel/preset-env",
+                        {
+                            "targets": {
+                                "node": "current"
+                            }
+                        }
+                    ]
+                ],
+                "plugins": [
+                    "istanbul"
+                ]
+            }
+        }
+    }
 }
