@@ -1,4 +1,4 @@
-import alert from './alert'
+import alert from './alert.js'
 export default alert
 
 
