@@ -1,6 +1,6 @@
-export * from './board';
-export * from './pieces';
-export * from './constants';
-export * from './rotations';
-export * from './types';
-export * from './rules';
+export * from './board.js';
+export * from './pieces.js';
+export * from './constants.js';
+export * from './rotations.js';
+export * from './types.js';
+export * from './rules.js';
